+++ Localization/deDE.lua
@@ -1,8 +1,8 @@
--- [[ Exported at 2022-11-07 22-47-55 ]] --
+-- [[ Exported at 2022-11-15 17-20-22 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --
 
--- [[ German translation by Ta, 2022-11-04 ]] --
+-- [[ German translation by Ta, 2022-11-14 ]] --
 
 -- [[ Namespaces ]] --
 local addonName, addon = ...;
@@ -224,16 +224,20 @@ L["will be"] = "wird";
 L["was"] = "war";
 L["only obtainable by one player"] = "nur von einem Spieler zu erzielen";
 L["Patch"] = "Patch";
-L["Temporary Obtainable Text"] = "{thisAchievement}{isWillBeWas}{neverOnceTempObt}{startText}{startDetail}{endText}{endDetail}.";
+L["Temporary Obtainable Text"] = "{thisAchievement}{isWillBeWas}{neverOnceTempObt}{startText}{startDetail}{endText}{endDetail}";
 L["Hide Not Earned By if current character earned the achievement"] = "{notEarnedBy} ausblenden, wenn der aktuelle Charakter den Erfolg schon erlangt hat";
 L["Hide Not Earned By if current character earned the achievement Desc"] = "Blendet den Abschnitt {notEarnedBy} aus, wenn der aktuelle Charakter den Erfolg erworben hat. Der Abschnitt {earnedBy} wird ansonsten weiterhin wie gewohnt angezeigt.";
 L["for achievement"] = "für {achievement}";
-L["Needs /love for achievement"] = "Benötigt deine /liebe für {achievement}";
-L["Already /loved for achievement"] = "Hat deine Liebe für {achievement} schon erhalten";
-L["Needs to be killed for achievement"] = "Muss noch für {achievement} getötet werden";
-L["Already killed for achievement"] = "Bereits getötet für {achievement}";
-L["Needs to be cooked for achievement"] = "Muss noch für {achievement} gekocht werden";
-L["Already cooked for achievement"] = "Bereits gekocht für {achievement}";
+L["Needs /love for achievement"] = "Benötigt deine /liebe {forAchievement}";
+L["Already /loved for achievement"] = "Hat deine Liebe {forAchievement} schon erhalten";
+L["Needs to be killed for achievement"] = "Muss noch {forAchievement} getötet werden";
+L["Already killed for achievement"] = "Bereits getötet {forAchievement}";
+L["Needs to be cooked for achievement"] = "Muss noch {forAchievement} gekocht werden";
+L["Already cooked for achievement"] = "Bereits gekocht {forAchievement}";
+L["Needs to be eaten for achievement"] = "Muss noch {forAchievement} gegessen werden";
+L["Already eaten for achievement"] = "Bereits gegessen {forAchievement}";
+L["Needs to be drunk for achievement"] = "Muss noch {forAchievement} getrunken werden";
+L["Already drank for achievement"] = "Bereits getrunken {for Achievement}";
 L["Data Manager"] = "Charakterübersicht";
 L["Header tooltip"] = "Überschrift-Tooltip";
 L["Ignore"] = "Ignorieren";
@@ -265,6 +269,8 @@ L["WoWInterface"] = "WoWInterface";
 L["WoWInterface Desc"] = "Öffnet ein Popup-Fenster mit einem Link zur Seite {addonName} {woWInterface}.";
 L["Show minimap icon"] = "Zeige Minimap Icon";
 L["Show minimap icon Desc"] = "Zeige / Verstecke das Minimap Icon.";
+L["Show world map icon"] = "Zeige Welt-Map Icon";
+L["Show world map icon Desc"] = "Zeige / Verstecke das Welt-Map Icon.";
 L["Reset view"] = "Ansicht zurücksetzen";
 L["Reset view on open"] = "Zurücksetzen der Ansicht beim Öffnen";
 L["Reset view on open Desc"] = "Alle Kategorien ausblenden und die erste Kategorie auswählen oder die zuletzt ausgewählte Kategorie beibehalten, wenn sie zuletzt geschlossen wurde.";