+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 46.4 - 2022-11-xx
+### Added
+- Tooltip data for items and recipes for Draenor Cuisine and Drown Your Sorrows
+- ItemRefTooltip now also shows tooltip data (the one you see when you click an item from chat)
+- Option to hide the World Map icon
+
+### Changed
+- Showing Tooltip Criteria if the achievement was not earned by the current character is checked by default
+- Tooltip Criteria is shown by default if the achievement is not completed
+
+### Fixed
+- Recipes should no longer show tooltip data more than once
+- A nil item should no longer cause any errors
+
+### Wrath Classic
+- Added some missing texts
+- Disabled the plugin to rename a dragonflight category there isn't
+
 ## 46.3 - 2022-11-09
 ### Added
 - Tooltip data for items and recipes for Second That Emotion, Cataclysmically Delicious, Pandaren Cuisine and Pandaren Delicacies